Extract shared placeholder-substitution algorithm into Command.build
rhizome/common/lib/util.rb's `cmd` and lib/net_ssh.rb's
NetSsh::WarnUnsafe.convert implemented the same :placeholder
shell-escaping algorithm independently. Move it into
rhizome/common/lib/command.rb as Command.build, required by both,
with `cmd` and WarnUnsafe.convert reduced to thin wrappers that pass
their own error-message label and strictness setting.

Command.build takes its own name/wrapper_file/strict as positional
arguments rather than keywords, so a legitimate placeholder keyword
(e.g. :name, as used by VmHostSliceNexus#prep) can never collide
with and get silently absorbed out of **kw.

Error messages locate the real caller via Thread.each_caller_location
(skipping frames in command.rb and the calling wrapper's file)
instead of a hardcoded caller(N, 1) depth, so it stays correct
regardless of how many wrapper layers exist between the call site
and this shared code. Falls back to omitting the location if that
method isn't available (added in Ruby 3.2; rhizome's CI still runs
3.0).

NetSsh::PotentialInsecurity and rhizome's top-level
PotentialInsecurity are now both aliases for
Command::PotentialInsecurity, so existing rescue/raise_error
references in both codebases keep working unchanged.
