+++ rhizome/common/lib/util.rb
@@ -6,6 +6,7 @@ require "bundler/setup" if File.directory?(File.expand_path("../../host", __dir_
 require "open3"
 require "shellwords"
 require "openssl"
+require_relative "command"
 
 class CommandFail < RuntimeError
   attr_reader :stdout, :stderr
@@ -26,76 +27,13 @@ class FsyncFail < Exception
 end
 # rubocop:enable Lint/InheritException
 
-class PotentialInsecurity < StandardError
-end
+PotentialInsecurity = Command::PotentialInsecurity
 
 # Safely build a shell command string from a template containing
 # :placeholder tokens, substituting each with the corresponding keyword
-# argument, shell-escaped. A keyword whose name starts with "shelljoin_"
-# is expected to hold an Array, joined into multiple shell-escaped words
-# instead of a single escaped word. Placeholders are not allowed inside
-# quotes, since shell-escaping a value for unquoted context and then
-# wrapping it in quotes anyway produces incorrect (and unsafe) results.
+# argument, shell-escaped. See Command.build for details.
 def cmd(command, **kw)
-  raise TypeError, "invalid type passed to cmd: #{command.inspect}" unless command.is_a?(String)
-  return command if kw.empty?
-  raise PotentialInsecurity, "Interpolated string passed to cmd at #{caller(1, 1).first}\nReplace interpolation with :placeholders and provide values for placeholders in keyword arguments." unless command.frozen?
-
-  result = +""
-  mode = :unquoted
-  base_re = Regexp.union(kw.keys.map(&:to_s))
-  unquoted_re, single_re, double_re = nil
-  until command.empty?
-    re = case mode
-    when :unquoted
-      unquoted_re ||= /(\\.|['"]|#.*$)|:(#{base_re})\b/
-    when :single
-      single_re ||= /(')|:(#{base_re})\b/
-    else # :double
-      double_re ||= /(\\.|")|:(#{base_re})\b/
-    end
-
-    pre, _, command = command.partition(re)
-    ch = $1
-    q = $2
-
-    if ch
-      case mode
-      when :unquoted
-        case ch
-        when "'"
-          mode = :single
-        when '"'
-          mode = :double
-        end
-      when :single
-        mode = :unquoted
-      else # :double
-        mode = :unquoted if ch == '"'
-      end
-      result << pre << ch
-    elsif mode != :unquoted
-      if q && !q.empty?
-        raise PotentialInsecurity, "Placeholder '#{q}' inside #{mode} quote in command at #{caller(1, 1).first}\nFix command to move the placeholder outside quotes, because shell escaping does not work correctly inside quotes."
-      end
-    else
-      result << pre
-      if q && !q.empty?
-        v = kw[q.to_sym]
-        result << if q.start_with?("shelljoin_")
-          v.shelljoin
-        else
-          v.to_s.shellescape
-        end
-      end
-    end
-  end
-
-  unless mode == :unquoted
-    raise PotentialInsecurity, "Unterminated #{mode} quote in command at #{caller(1, 1).first}\nFix command syntax."
-  end
-
-  result.freeze
+  Command.build(command, "cmd", __FILE__, true, **kw)
 end
 
 # :nocov:
